Refactor container package for IT (#9822)
* Refactor container package for IT

* Remove useless log
