+++ shardingsphere-test/shardingsphere-integration-test/shardingsphere-integration-test-fixture/src/test/java/org/apache/shardingsphere/test/integration/junit/container/adapter/ShardingSphereAdapterContainer.java
@@ -15,12 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.test.integration.junit.container;
+package org.apache.shardingsphere.test.integration.junit.container.adapter;
 
 import lombok.Getter;
 import lombok.Setter;
-import lombok.extern.slf4j.Slf4j;
 import org.apache.shardingsphere.test.integration.junit.annotation.XmlResource;
+import org.apache.shardingsphere.test.integration.junit.container.ShardingSphereContainer;
 import org.apache.shardingsphere.test.integration.junit.processor.AuthenticationProcessor;
 import org.apache.shardingsphere.test.integration.junit.processor.AuthenticationProcessor.Authentication;
 
@@ -29,7 +29,6 @@ import javax.sql.DataSource;
 /**
  * ShardingSphere adapter container.
  */
-@Slf4j
 public abstract class ShardingSphereAdapterContainer extends ShardingSphereContainer {
     
     @Getter
